refactor/refactor LabResultComponent to atom
